+++ Main/src/main/java/services/QuadraticEquationRootsFinder.java
@@ -1,4 +1,4 @@
-package models;
+package services;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -11,9 +11,9 @@ public class QuadraticEquationRootsFinder {
 
     private static final Logger logger = LoggerFactory.getLogger(QuadraticEquationRootsFinder.class);
 
-    private static String decision;
+    private String decision;
 
-    public static void run(double firstArr, double secondArr, double thirdArr) {
+    public void run(double firstArr, double secondArr, double thirdArr) {
         double disc = getDescriminant(firstArr, secondArr, thirdArr);
         logger.info(String.format("\nDiscriminant = %s", disc));
         if (disc > 0) {
@@ -23,27 +23,28 @@ public class QuadraticEquationRootsFinder {
         } else if (disc < 0) {
             decision = getNoDecisions();
         }
+        logger.info("\nSo your decision is: {}", decision);
     }
 
     public String getDecision() {
         return decision;
     }
 
-    private static double getDescriminant(double firstArr, double secondArr, double thirdArr) {
+    private double getDescriminant(double firstArr, double secondArr, double thirdArr) {
         return (secondArr * secondArr) - (4 * (firstArr * thirdArr));
     }
 
-    private static String getFullDecision(double firstArr, double secondArr, double thirdArr) {
+    private String getFullDecision(double firstArr, double secondArr, double thirdArr) {
         double x1 = ((-secondArr + sqrt(getDescriminant(firstArr, secondArr, thirdArr))) / (firstArr * 2));
         double x2 = ((-secondArr - sqrt(getDescriminant(firstArr, secondArr, thirdArr))) / (firstArr * 2));
         return Arrays.toString(new double[]{x1, x2});
     }
 
-    private static double getOneDecision(double firstArr, double secondArr) {
+    private double getOneDecision(double firstArr, double secondArr) {
         return (secondArr * -1) / (2 * firstArr);
     }
 
-    private static String getNoDecisions() {
+    private String getNoDecisions() {
         return ("Discriminant < 0, there's no decisions");
     }
 }
