+++ Main/src/main/java/views/ExampleAppEquationView.java
@@ -4,9 +4,7 @@ import controllers.Controller;
 public class ExampleAppEquationView {
 
     public void startApp(String[] args) {
-        Controller contr = new Controller();
-        contr.run(args);
-        String viewVariable = contr.getD();
-//        System.out.println(viewVariable);
+        Controller contr = new Controller(args);
+        contr.getDecision();
     }
 }
